+++ src/Controller/Api/PictureController.php
@@ -4,6 +4,7 @@ namespace App\Controller\Api;
 
 use App\Entity\Memory;
 use App\Entity\Picture;
+use App\Service\FileUploader;
 use OpenApi\Attributes as OA;
 use App\Repository\PlaceRepository;
 use App\Repository\MemoryRepository;
@@ -22,11 +23,21 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
  * A first one displays all additional pictures.
  * A second one displays a single additional picture by its id.
  * A third one uploads or updates the main memory picture.
- * A fourth one uploads or updates (an) addditional memory picture(s).
- * A fifth and last one deletes a additional picture by its id.
+ * A fourth one uploads (an) addditional memory picture(s).
+ * A fourth one updates an addditional memory picture.
+ * A sixth and last one deletes a additional picture by its id.
  */
 class PictureController extends AbstractController
 {
+
+    private $fileUploader;
+
+    public function __construct(FileUploader $fileUploader)
+    {
+        $this->fileUploader = $fileUploader;
+    }
+
+
     /**
      * Display all additional pictures
      * @param PictureRepository $ppictureRepository
@@ -104,13 +115,8 @@ class PictureController extends AbstractController
                 404
             );
         }
+        return $this->json(['picture' => $picture], Response::HTTP_OK, [], ['groups' => ['get_picture', 'get_memory_id']]);
 
-        return $this->json(
-            $picture,
-            200,
-            [],
-            ['groups' => ['get_picture', 'get_memory_id']]
-        );
     }
 
     /**
@@ -200,15 +206,15 @@ class PictureController extends AbstractController
         }
         // Continue with the code to handle the case when a new picture is uploaded
 
-        // on ajoute uniqid() afin de ne pas avoir 2 fichiers avec le même nom
-        $newFilename = uniqid() . '.' . $picture->getClientOriginalExtension();
-
-        // enregistrement de l'image dans le dossier public du serveur
-        // params->get('public') =>  va chercher dans services.yaml la variable public
-        $picture->move($params->get('images_directory'), $newFilename);
-
-        // ne pas oublier d'ajouter l'url de l'image dans l'entité appropriée
-        // $entity est l'entity qui doit recevoir votre image
+        if ($memory->getMainPicture()) {
+            $deleteFileResult = $this->fileUploader->deletePictureFile($params->get('images_directory'), $memory->getMainPicture());
+            if (!$deleteFileResult) {
+            return $this->json("Erreur : Échec de suppression de la photo", 500);
+            }
+        }
+        
+        $newFilename = $this->fileUploader->uploadImage($picture);
+        
         $memory->setMainPicture($newFilename);
         $entityManager->flush();
 
@@ -216,17 +222,139 @@ class PictureController extends AbstractController
     }
 
     /**
-     * Upload or update (an/the) additional memory picture(s)
-     * 
-     * @param
+     * Upload (an/the) additional memory picture(s)
+     * @param Memory $memory
+     * @param Request $request
+     * @param ParameterBagInterface $params
+     * @param EntityManagerInterface $entityManager
      * @return Response
      */
+    #[OA\Parameter(
+        name: "id",
+        in: "path",
+        required: true,
+        description: "ID of the memory",
+        schema: new OA\Schema(type: 'integer')
+    )]
+    #[OA\RequestBody(
+        description: 'Exemple of data to be supplied to upload the additional picture',
+        content: new OA\JsonContent(
+            properties: [
+                new OA\Property(property: 'picture', type: 'file', example: 'photo.jpg'),
+                new OA\Property(
+                    property: "memory",
+                    type: "object",
+                    properties: [
+                        new OA\Property(property: "id", type: "integer", example: 9)
+                    ]
+                )
+            ]
+        )
+    )]
+    #[OA\Response(
+        response: 201,
+        description: 'Saves the image associated with the memory',
+        content: new OA\JsonContent(
+            type: 'array',
+            items: new OA\Items(ref: new Model(type: Memory::class, groups: ['get_memory', 'get_location', 'get_picture', 'get_place', 'get_user']))
+        )
+    )]
+    #[OA\Tag(name: 'picture')]
     #[Route('api/secure/upload/additional_pictures/{id<\d+>}', methods: ['POST'])]
-    public function upload_update_additional_pictures(): Response
+    public function upload_additional_pictures(Memory $memory, Request $request, ParameterBagInterface $params, EntityManagerInterface $entityManager): Response
     {
-        return $this->json([
-            'message' => 'Image(s) supplémentaire(s) téléchargée(s) et associée(s) au souvenir avec succès.'
-        ]);
+
+         /** @var \App\Entity\User $user */
+         $user = $this->getUser();
+
+         if ($user !== $memory->getUser()) {
+             return $this->json("Erreur : Vous n'êtes pas autorisé à ajouter de photo à ce souvenir.", 401);
+         }
+         // Retrieve the uploaded picture file from the request
+         $pictures = $request->files->get('additional_pictures');
+
+        $newPictures = [];
+
+        foreach ($pictures as $picture) {
+
+        $newFilename = $this->fileUploader->uploadImage($picture);
+
+        // ne pas oublier d'ajouter l'url de l'image dans l'entité appropriée
+        // $entity est l'entity qui doit recevoir votre image
+        $newPicture = (new Picture())
+        ->setPicture($newFilename)
+        ->setMemory($memory);
+        $entityManager->persist($newPicture);
+
+        $newPictures[]= $newPicture;
+        }
+        $entityManager->flush();
+
+
+        return $this->json(['pictures' => $newPictures, 'message' => 'Image(s) téléchargée(s) et associée(s) au souvenir avec succès.'], Response::HTTP_CREATED, [], ['groups' => ['get_memory', 'get_location', 'get_place', 'get_user', 'get_picture']]);
+    }
+
+      /**
+     * Update an additional memory picture
+     * @param Picture $picture
+     * @param Request $request
+     * @param ParameterBagInterface $params
+     * @param EntityManagerInterface $entityManager
+     * @return Response
+     */
+    #[OA\Parameter(
+        name: "id",
+        in: "path",
+        required: true,
+        description: "ID of the picture",
+        schema: new OA\Schema(type: 'integer')
+    )]
+    #[OA\RequestBody(
+        description: 'Exemple of data to be supplied to update the additional picture',
+        content: new OA\JsonContent(
+            properties: [
+                new OA\Property(property: "id", type: "integer", example: 9), 
+                new OA\Property(property: 'picture', type: 'file', example: 'photo.jpg'),
+            ]
+        )
+    )]
+    #[OA\Response(
+        response: 201,
+        description: 'Saves the image associated with the memory',
+        content: new OA\JsonContent(
+            type: 'array',
+            items: new OA\Items(ref: new Model(type: Memory::class, groups: ['get_memory', 'get_location', 'get_picture', 'get_place', 'get_user']))
+        )
+    )]
+    #[OA\Tag(name: 'picture')]
+    #[Route('api/secure/update/additional_pictures/{id<\d+>}', methods: ['POST'])]
+    public function update_additional_pictures(Picture $picture, Request $request, ParameterBagInterface $params, EntityManagerInterface $entityManager): Response
+    {
+
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+
+        if ($user !== $picture->getMemory()->getUser()) {
+             return $this->json("Erreur : Vous n'êtes pas autorisé à ajouter de photo à ce souvenir.", 401);
+        }
+         
+        $newPicture = $request->files->get('additional_pictures');
+
+        if ($newPicture === null) {
+            $picture->getPicture();
+            $entityManager->flush();
+            return $this->json(['message' => 'Le souvenir a bien été mis à jour.']);
+            }
+        $currentPictureDeleteResult = $this->fileUploader->deletePictureFile($params->get('images_directory'), $picture->getPicture());
+        if (!$currentPictureDeleteResult) {
+            return $this->json("Erreur : Échec de suppression de l'ancienne photo", 500);
+        }
+
+        $newFilename = $this->fileUploader->uploadImage($newPicture);
+        $picture->setPicture($newFilename);
+        $entityManager->flush();
+
+        return $this->json(['picture' => $picture, 'message' => 'Image mise à jour avec succès.'], Response::HTTP_CREATED, [], ['groups' => ['get_memory', 'get_location', 'get_place', 'get_user', 'get_picture']]);
     }
 
     /**
@@ -251,7 +379,7 @@ class PictureController extends AbstractController
         schema: new OA\Schema(type: 'integer')
     )]
     #[OA\Tag(name: 'picture')]
-    public function delete(Picture $picture, EntityManagerInterface $entityManager, MemoryRepository $memoryRepository): Response
+    public function delete(Picture $picture, EntityManagerInterface $entityManager, MemoryRepository $memoryRepository, ParameterBagInterface $params, ): Response
     {
         /** @var \App\Entity\User $user */
         $user = $this->getUser();
@@ -267,6 +395,12 @@ class PictureController extends AbstractController
                 "Erreur : La photo n'existe pas",
                 404
             );
+        }
+
+        $deleteFileResult = $this->fileUploader->deletePictureFile($params->get('images_directory'), $picture->getPicture());
+
+        if (!$deleteFileResult) {
+            return $this->json("Erreur : Échec de suppression de la photo", 500);
         }
         $entityManager->remove($picture);
         $entityManager->flush();
